backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
mlflow==2.11.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
torchvision==0.17.1+cpu
efficientnet-pytorch==0.7.1
facenet-pytorch==2.5.3
opencv-python-headless==4.9.0.80
numpy==1.26.4
prometheus-client==0.20.0
python-json-logger==2.0.7
httpx==0.27.0
Pillow==10.2.0
